be/ai_module/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_module"
version = "0.1.0"
description = "AI Module for Capstone Project"
dependencies = [
    "pika",
    "python-dotenv"
]
